Move clock service into addon folder
This makes it possible to have multiple separate clock services with e.g. different interval times
